requirements.txt
pygame-ce==2.4.0
numpy==1.26.2
mediapipe==0.10.11